requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.15.4
Flask==3.0.3
Flask-RESTful==0.3.10
gitdb==4.0.11
GitPython==3.1.41
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.0.1
openai==1.37.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
